+++ ydb/core/fq/libs/row_dispatcher/format_handler/filters/ya.make
@@ -9,9 +9,9 @@ PEERDIR(
     ydb/core/fq/libs/actors/logging
     ydb/core/fq/libs/row_dispatcher/events
     ydb/core/fq/libs/row_dispatcher/format_handler/common
-    ydb/core/fq/libs/row_dispatcher/purecalc_no_pg_wrapper
 
     ydb/library/actors/core
+    ydb/library/purecalc/no_pg_wrapper
 
     yql/essentials/minikql
     yql/essentials/minikql/computation
