+++ ydb/library/purecalc/ya.make
@@ -1,0 +1,4 @@
+RECURSE(
+    compiliation
+    no_pg_wrapper
+)
